(Feat) Added new page for specifying server ip to connect to. (Refactor) Changed project name.
