+++ Game.xaml
@@ -1,0 +1,17 @@
+﻿<Page x:Class="WpfApp1.Game"
+      xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+      xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+      xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
+      xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
+      xmlns:local="clr-namespace:WpfApp1"
+      mc:Ignorable="d" 
+      d:DesignHeight="450" d:DesignWidth="800"
+      Title="Game">
+
+    <Grid>
+        <Canvas x:Name="GameCanvas" Background="White">
+            <!-- Код для отрисовки поля и игроков генерируется программно -->
+
+        </Canvas>
+    </Grid>
+</Page>
