+++ MainWindow.xaml
@@ -1,9 +1,8 @@
-﻿<Window x:Class="YourNamespace.MainWindow"
+﻿<Window x:Class="WpfApp1.MainWindow"
         xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
         xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
         Title="Bomberman" Height="600" Width="650">
-    <Canvas x:Name="GameCanvas" Background="White">
-        <!-- Код для отрисовки поля и игроков генерируется программно -->
-            
-    </Canvas>
+    <Grid>
+        <Frame x:Name="Main" NavigationUIVisibility="Hidden"/>
+    </Grid>
 </Window>
