+++ Menu.xaml
@@ -1,0 +1,27 @@
+﻿<Page x:Class="WpfApp1.Menu"
+      xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+      xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+      xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
+      xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
+      xmlns:local="clr-namespace:WpfApp1"
+      mc:Ignorable="d" 
+      d:DesignHeight="450" d:DesignWidth="800"
+      Title="Menu">
+
+    <Grid Grid.Column="1" VerticalAlignment="Center" HorizontalAlignment="Center"> 
+        <Grid.RowDefinitions>
+            <RowDefinition/>
+            <RowDefinition/>
+            <RowDefinition/>
+        </Grid.RowDefinitions>
+
+        <Label Padding="10" FontWeight="DemiBold" Grid.Row="0" FontFamily="Montserrat" FontSize="15" Content="Server IP"></Label>
+        <TextBox x:Name="ServerIP" VerticalContentAlignment="Center" FontFamily="Inter" Text="" TextAlignment="Center" FontSize="15" Grid.Row="1" Height="30" Width="150"/>
+
+        <Button x:Name="Run_Button" Click="Button_Click" Grid.Row="2" FontFamily="Montserrat" FontSize="15" Content="Connect" VerticalAlignment="Center" HorizontalAlignment="Center" Width="90" Height="30" BorderBrush="#FFDDDDDD" Cursor="Hand" Background="#FFDDDDDD">
+            <Button.Effect>
+                <DropShadowEffect ShadowDepth="1" BlurRadius="10" Opacity="0.1"/>
+            </Button.Effect>
+        </Button>
+    </Grid>
+</Page>
